+++ HISTORY.md
@@ -11,6 +11,7 @@
 ### 2022-03-09
 * new dev version
 * byob, store, gui: export / import sprite-local custom block definitions, under construction
+* byob, gui: adapted library import dialog to the new format
 
 ## 7.3.0:
 * **New Features:**